pronin_test_task/requirements.txt
djangorestframework==3.14
python-dotenv==0.21.1
psycopg2-binary==2.9.6
gunicorn==20.0.4
requests==2.31.0